BIOME-0100 - added the Firebase dependencies to project
