+++ .idea/workspace.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ChangeListManager">
+    <list default="true" id="e688ac12-d362-4bcf-8cf0-36dde0b446b9" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/BiomeOfus.xcodeproj/project.xcworkspace/xcuserdata/sergionunez.xcuserdatad/IDEFindNavigatorScopes.plist" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/Forest.imageset/Contents.json" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/Forest.imageset/Forest.png" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/appleTrees.imageset/Contents.json" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/appleTrees.imageset/appleTrees.jpeg" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/theAppleTree.imageset/Contents.json" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/Assets.xcassets/theAppleTree.imageset/theAppleTree.jpeg" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/BiomeForm.swift" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/BiomeView.swift" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/BiomeOfus/DashboardView.swift" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus.xcodeproj/project.pbxproj" beforeDir="false" afterPath="$PROJECT_DIR$/BiomeOfus.xcodeproj/project.pbxproj" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus.xcodeproj/project.xcworkspace/xcuserdata/sergionunez.xcuserdatad/UserInterfaceState.xcuserstate" beforeDir="false" afterPath="$PROJECT_DIR$/BiomeOfus.xcodeproj/project.xcworkspace/xcuserdata/sergionunez.xcuserdatad/UserInterfaceState.xcuserstate" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus/BiomeOfusApp.swift" beforeDir="false" afterPath="$PROJECT_DIR$/BiomeOfus/BiomeOfusApp.swift" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus/ContentView.swift" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus/LoginScn.swift" beforeDir="false" afterPath="$PROJECT_DIR$/BiomeOfus/LoginScn.swift" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/BiomeOfus/ProfileView.swift" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="GitSEFilterConfiguration">
+    <file-type-list>
+      <filtered-out-file-type name="LOCAL_BRANCH" />
+      <filtered-out-file-type name="REMOTE_BRANCH" />
+      <filtered-out-file-type name="TAG" />
+      <filtered-out-file-type name="COMMIT_BY_MESSAGE" />
+    </file-type-list>
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2fNFd1LhKF5c6yPhmXbx6ofjTgg" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">
+    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
+    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="e688ac12-d362-4bcf-8cf0-36dde0b446b9" name="Changes" comment="" />
+      <created>1713633711622</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1713633711622</updated>
+      <workItem from="1713633718797" duration="7000" />
+    </task>
+    <servers />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+</project>
