+++ BiomeOfus/Extensions/String.swift
@@ -1,0 +1,16 @@
+//
+//  Validator.swift
+//  BiomeOfus
+//
+//  Created by Jaresi Nunez on 4/21/24.
+//
+
+import Foundation
+
+extension String {
+    func isValidEmail() -> Bool {
+        let regex = try! NSRegularExpression(pattern: "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}", options: .caseInsensitive)
+        return regex.firstMatch(in: self, range: NSRange(location: 0, length: count)) != nil
+    }
+}
+
